fix: delegation approver functions
